mk: add coverage-target
xNVMe has practices to create a command-line tool and possible also an
explicit test-tool for each function exposed in the public API, the
tools are "parameterizable" for the logic of the tool/API and for
backend instrumentation. These practices ensure that the API is tested
against the different backends.

However, the "coverage" is not reported/automated, this adds a
short-hand to the toolback to build with "coverage" e.g. gcov support.
Thus, after running the test-workflows, then the following can be
invoked:

gcovr --html /tmp/report.html

Showing a coverage report. This serves as a starting point for
automatically documentating the code-coverage of the testplans.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>