feat(toolbox/pkgs): replace SPDK subproject with 3p
This replaces the consumption of SPDK via subproject, to be consumed
from user-provided path to SPDK repository, and assuming that when
building xNVMe, then SPDK has already been built.

The way that the SPDK subproject is currently utilized is a
subversion/hack of Meson functionality. The meson setup/configure stage
is hi-jacked to be download the SPDK subproject and at configure-time,
go into the subproject to configure and build SPDK. This is done since
SPDK is not a meson-project, and thus cannot integrate properly in the
required meson-stages for configuration/build etc.

The changes are wide and involves:

* Ensure license descriptions in REUSE

* Migrate scripts from subproject-package files to
  - Scripts in toolbox/pkgs/*.sh, describing how to retrieve, configure
    and build SPDK
  - Templates are used for this, which live in
    toolbox/pkgs/emitter/templates/*
  - Move patches from the subrproject folder into the third-party folder

* Adjust the main meson.build

  - Change the meson-option "with-spdk" from a boolean to a "string".
    This is to carry the path to the SPDK repository
  - Add a "toolbox/third-party/spdk/meson.build" encapsulating all the
    SPDK concerns

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>