refactor(api/be)!: remove xnvme_be_attr_list
The original intent of the attribute list was to allow library users to
programmatically inspect the backend capabilities of the library.

However, the current API for this purpose is not ideal. Rather than
retaining a subpar implementation in the public API, it has been
removed. Its usage in tools such as CLI utilities remains valuable, so a
simple information printer will serve as a replacement.

This change is part of the preparation for improving the API for code
generation and the eventual v1 release.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>